pyproject.toml
[project]
name = "javarush-django-project"
version = "0.1.0"
description = "Python University Module 3 - Django Final Project"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
]
